Make a hover effect over the main image including the view-icon
Edited the main image to have a hover effect that turns a specific color
with the view-icon showing in the center. I made the anchor tag with the
position reletive and the image with the position absolute. I have never
used these pseudo-classes before, including some of the backgorund
properties, and ::before & ::after. After doing some research on these
topics I was able to acheive this effect.